requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
PyYAML>=6.0
Pillow>=10.0,<11.0
